+++ earth2studio/lexicon/__init__.py
@@ -37,6 +37,7 @@ from .goes_glm import GOESGLMLexicon
 from .himawari_ahi import HimawariAHILexicon
 from .hrrr import HRRRFXLexicon, HRRRLexicon
 from .ibtracs import IBTrACSLexicon
+from .iem import IEM_ASOSLexicon
 from .isd import ISDLexicon
 from .jpss import JPSSATMSLexicon, JPSSCrISLexicon, JPSSLexicon
 from .meteosat import MeteosatFCILexicon